Fix PHP 8.5 deprecations
- GSObject::getDouble(): replace the non-canonical (double) cast with
  (float). PHP 8.5 deprecates alias casts ("Non-canonical cast (double)
  is deprecated"); the deprecation is emitted at compile time whenever
  the file is loaded. (float) is an exact alias, identical behavior on
  every PHP version.

- GSRequest::sendRequest(): drop the curl_close() call. Since PHP 8.0
  curl handles are objects and curl_close() has no effect; PHP 8.5
  deprecates calling it. The handle is freed when it goes out of scope.
  The SDK already requires PHP >= 8.0, so this is a no-op removal.

+++ src/GSObject.php
@@ -153,7 +153,7 @@ class GSObject
 	 */
     public function getDouble($key, $defaultValue = GSObject::DEFAULT_VALUE)
     {
-        return (double)$this->get($key, $defaultValue);
+        return (float)$this->get($key, $defaultValue);
     }
 
 	/**
